resonance revamp, clickable team char
